requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
curl_cffi==0.14.0
et_xmlfile==2.0.0
fredapi==0.5.2
frozendict==2.4.7
holidays==0.57
html5lib==1.1
idna==3.11
joblib==1.5.3
lxml==6.0.2
multitasking==0.0.12
numpy==1.26.4
openpyxl==3.1.5
packaging==26.0
pandas==2.2.3
patsy==1.0.2
peewee==4.0.0
platformdirs==4.9.2
protobuf==7.34.0
pyarrow==17.0.0
pycparser==3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.17.0
soupsieve==2.8.3
statsmodels==0.14.4
ta==0.11.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webencodings==0.5.1
websockets==16.0
yfinance==0.2.66
